requirements.txt
streamlit
opencv-python
numpy
Pillow
ultralytics
streamlit-carousel